# docs/openapi/prima-pool.yaml
openapi: 3.1.0
info:
  title: prima-pool control plane API
  version: 0.3.0
  description: >
    Negotiation protocol for the prima-pool inference pool. Accounts register with a username and
    password, create scoped API keys (worker / user), and worker devices log in with a worker key
    to join WireGuard clusters. Usage/accounting endpoints are intentionally out of scope
    (see docs/protocol/usage-and-accounting.md).
servers:
  - url: https://pool.example.com/v1
...
paths:
  /accounts/register:
    post:
      tags: [accounts]
      operationId: registerAccount
      summary: Register an account (username + password).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAccountRequest"
      responses:
        "201":
          description: Account registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /accounts/login:
    post:
      tags: [accounts]
      operationId: login
      summary: Log in and receive a short-lived session token for key management.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /accounts/{account_id}/keys:
    post:
      tags: [accounts]
      operationId: createKey
      summary: Create a scoped API key (worker or user).
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeyRequest"
      responses:
        "201":
          description: Key created (secret shown once).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [accounts]
      operationId: listKeys
      summary: List the account's API keys (without secrets).
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: List of keys.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKeySummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /accounts/{account_id}/dashboard:
    get:
      tags: [accounts]
      operationId: accountDashboard
      summary: Account-scoped dashboard data (workers + keys).
      description: >
        Used by the static GUI at /ui. Returns the account's workers and API
        keys. Auth: the account's session token (from /accounts/login).
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Dashboard data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id: { type: string }
                  username: { type: string }
                  workers:
                    type: array
                    items:
                      type: object
                      properties:
                        worker_id: { type: string }
                        model: { type: string }
                        status:
                          type: string
                          enum: [registered, waitlisted, assigned]
                        online: { type: boolean }
                        memory_mb: { type: integer }
                        cluster_id: { type: string }
                  keys:
                    type: array
                    items:
                      type: object
                      properties:
                        key_id: { type: string }
                        name: { type: string }
                        scope: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /accounts/{account_id}/keys/{key_id}:
    delete:
      tags: [accounts]
      operationId: revokeKey
      summary: Revoke a scoped API key.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/AccountId"
        - $ref: "#/components/parameters/KeyId"
      responses:
        "204":
          description: Key revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /workers/register:
    post:
      tags: [workers]
      operationId: registerWorker
      summary: Register a worker device with a worker-scoped key (joins the model's waitlist).
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterWorkerRequest"
      responses:
        "201":
          description: Worker registered, waitlisted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /workers/{worker_id}:
    delete:
      tags: [workers]
      operationId: revokeWorker
      summary: Revoke a worker (leave the waitlist / cluster).
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkerId"
      responses:
        "204":
          description: Worker revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /workers/{worker_id}/state:
    get:
      tags: [workers]
      operationId: getWorkerState
      summary: Recover the worker's full current state (WS recovery path).
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkerId"
      responses:
        "200":
          description: Worker state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /workers/{worker_id}/heartbeat:
    post:
      tags: [workers]
      operationId: heartbeat
      summary: Report liveness.
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkerId"
      responses:
        "200":
          description: Heartbeat acknowledged.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /clusters/{cluster_id}/config:
    get:
      tags: [clusters]
      operationId: getClusterConfig
      summary: Fetch the WireGuard cluster config (idempotent, re-pullable).
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/ClusterId"
      responses:
        "200":
          description: Cluster config.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /chat/completions:
    post:
      tags: [inference]
      operationId: chatCompletions
      summary: Proxy an OpenAI-compatible chat completion to a live cluster's head.
      description: >
        Auth: user-scoped API key. The server finds a live cluster for the requested
        model and forwards the request to the head's llama-server over the WireGuard
        tunnel (option A: the server joins the cluster WG network). Streaming (SSE)
        is supported when `stream: true`.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, messages]
              properties:
                model: { type: string }
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role: { type: string }
                      content: { type: string }
                stream: { type: boolean, default: false }
                max_tokens: { type: integer }
      responses:
        "200":
          description: Chat completion (or SSE stream).
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Upstream cluster head unreachable.
components:
  securitySchemes:
    apiKeyAuth:
      type: http
      scheme: bearer
      description: Scoped API key (worker or user), e.g. `sk-worker-abc123`.
    sessionAuth:
      type: http
      scheme: bearer
      description: >
        Short-lived session token from /accounts/login, e.g. `sess_...`.
        On worker endpoints (`/workers/...`, `/clusters/.../config`,
        `/clusters/.../ready`, the WS channel), EITHER the worker-scoped API
        key OR the owning account's session token is accepted (both must own
        the target worker).
  parameters:
    AccountId:
      name: account_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^acc_[A-Za-z0-9]+$"
      description: Account ID returned at registration.
    WorkerId:
      name: worker_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^wrk_[A-Za-z0-9]+$"
    KeyId:
      name: key_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^key_[A-Za-z0-9]+$"
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Invalid or missing credentials.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated but not allowed (e.g. not a cluster member).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Conflict:
      description: State conflict (e.g. username taken).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    RegisterAccountRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    Account:
      type: object
      required: [account_id, username, created_at]
      properties:
        account_id: { type: string }
        username: { type: string }
        created_at:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    Session:
      type: object
      required: [session_token, expires_at]
      properties:
        session_token: { type: string }
        expires_at:
          type: string
          format: date-time
    CreateKeyRequest:
      type: object
      required: [name, scope]
      properties:
        name: { type: string }
        scope:
          type: string
          enum: [worker, user]
    ApiKey:
      type: object
      required: [key_id, name, scope, api_key, created_at]
      properties:
        key_id: { type: string }
        name: { type: string }
        scope:
          type: string
          enum: [worker, user]
        api_key:
          type: string
          description: Shown only once, at creation.
        created_at:
          type: string
          format: date-time
    ApiKeySummary:
      type: object
      required: [key_id, name, scope, created_at]
      properties:
        key_id: { type: string }
        name: { type: string }
        scope:
          type: string
          enum: [worker, user]
        created_at:
          type: string
          format: date-time
...
    RegisterWorkerRequest:
      type: object
      required: [model, gguf_sha256, memory_allocated_mb, wg_pubkey, endpoint]
      properties:
        model: { type: string }
        gguf_sha256:
          type: string
          minLength: 64
          maxLength: 64
          description: >
            SHA-256 of the worker's GGUF file. The server only groups workers
            whose hash matches the registered model's hash — a mismatched GGUF
            is rejected, so a cluster always runs the same model with the same
            quantization.
        memory_allocated_mb: { type: integer, minimum: 1 }
        wg_pubkey:
          type: string
          description: Device's WireGuard public key (private key never leaves the device).
        endpoint:
          $ref: "#/components/schemas/EndpointInfo"
        hardware:
          $ref: "#/components/schemas/Hardware"
...
    WorkerState:
      type: object
      required: [worker_id, account_id, status, model]
      properties:
        worker_id: { type: string }
        account_id: { type: string }
        status:
          type: string
          enum: [registered, waitlisted, assigned]
          description: >
            Same enum as Worker.status; `ready`/`active` are not worker states. Liveness is
            `online`; cluster readiness is the cluster's own status.
        online:
          type: boolean
          description: Liveness flag; false means the worker is down (out of the waitlist).
        model: { type: string }
        cluster:
          $ref: "#/components/schemas/ClusterAssignment"
...